check_apt_updates: Ignore new packages from upgrade list
Add extra check for discarding packages which are not currently installed, but instead are to be installed as a dependency of another upgrade.

In some situations this will cause the newly installed package (dependency) to come from the "security" repo and thus be considered as a critical upgrade.

Furthermore, a package that is not currently installed should never be considered as "upgrade" and should not be reported by this check as "pending upgrade".

Additionally fix error with function to generate the apt cache without any progress echo on stdout